server/server/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
distlib==0.3.9
Django==5.1.3
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
filelock==3.16.1
frozenlist==1.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
multidict==6.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.2
twilio==9.3.7
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
yarl==1.18.0
zope.interface==7.1.1
